requirements.txt
Flask==2.3.3
boto3==1.28.85
python-dotenv==1.0.0
Werkzeug==2.3.7
pytest==7.4.3
pytest-cov==4.1.0
mcp==0.9.0